+++ conans/model/recipe_ref.py
@@ -181,7 +181,7 @@ class RecipeReference:
             # timestamp
             tokens = rref.rsplit("%", 1)
             text = tokens[0]
-            timestamp = int(tokens[1]) if len(tokens) == 2 else None
+            timestamp = float(tokens[1]) if len(tokens) == 2 else None
 
             # revision
             tokens = text.split("#", 1)
